feat: setup API keys in global conftest.py
Without this change, tests like test_responses_client.py will pass when
run as part of a full test suite invocation (because the e2e tests run
first, and setup the API keys). However, they fail when run in isolation
(because they require valid API keys that aren't set.)
